[plugins] Added org.sailfishos.nfc.Tag.Removed signal
It replaces PresentChanged signal which was never issued in the
first place (DBusServiceTag was getting deleted before it had a
chance to issue this signal), so this can't really be considered
an API break. Still, the interface version is bumped, so that
support for the new signal could be determined at runtime by
those clients which depend on it.

The new Removed signal is always issued when DBusServiceTag is
deleted. This way, D-Bus client won't have to watch the adapter
too in order to be notified when the tag is gone.