[core] Serialize reactivation and transmit requests. JB#51200
It didn't matter before, because reactivation was only happening during
initialization of ISO-DEP targets. The whole process was controlled by
nfcd, no external requests could be submitted until after the target
has been initialized, everything was nice and simple.

Now that we are considering exposing reactivation functionality over
D-Bus, we need to serialize reactivation and transmit requests and
that's a totally different game.