[nfcd] Expose more poll parameters. JB#51099
Now it is possible to detect ID(UID) of the tags using poll parameters:
NFCID0 - NFC-B identifier of the NFC Forum Device. 4 bytes.
NFCID1 - NFC-A identifier of the NFC Forum Device in the Passive
Communication mode. Can be 4, 7, or 10 bytes.

These values can be used to identify tags (even though the value could be random).