[tools] Made ndef-share work more reliably
Deactivating the interface right after the final response has been
sent over NFC may be too early for the other side - it may still need
some time to fully process our response.

Better to wait either for reactivation of the interface initiated
by the other side, or the end of the touch.